Got rid of msmtpq for now since it gave misleading output
I found that if one email bounced it got added to the queue and then
the next time I tried to send an email I'd get output in the
minibuffer back about the first email again. That mislead me to think
that the wrong email server was being used and that my emails were
bouncing. I'm going to try nullmailer next instead!