Disable tag/folder mapping for work account
It turns out that the tag/folder mapping gets confused when forwarding
emails between two accounts that have tag/folder mapping enabled. That
is since the email will always be moved into the
first (alphabetically) account's directory on sync and that screws up
the tags too.

I don't have time to rethink that for now, so let's just disable it
for my work account and update the README.