Output "Waiting..." between syncs
It can be hard to know when post-processing has finished, so let's add
a "Waiting..." message to make it clear. In the past, I would hit
Enter to start mbsync again, before killing it. I found out the hard
way that is risky, since the partial sync state is saved in lockfiles
and can be corrupted on upgrade.